requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
mysqlclient==2.2.1
python-decouple==3.8
Pillow==10.2.0
celery==5.3.6
redis==5.0.1
requests==2.31.0
PyJWT==2.8.1
cryptography==42.0.2
